[MOV] test_orm: test_attributes into test_fields
