+++ odoo/addons/test_orm/tests/__init__.py
@@ -2,7 +2,6 @@ from . import (
     test_access_feedback,
     test_access_monetary_related,
     test_action_bindings,
-    test_attributes,
     test_autovacuum,
     test_check_access,
     test_company_checks,
