+++ odoo/addons/test_orm/tests/test_fields.py
@@ -5253,3 +5253,14 @@ class TestModifiedPerformance(TransactionCase):
         # One INSERT
         with self.assertQueryCount(1):
             self.ModifiedLine.create({})
+
+
+@tagged('at_install', '-post_install')
+class TestAttributes(TransactionCase):
+
+    def test_we_cannot_add_attributes(self):
+        Model = self.env['test_orm.category']
+        instance = Model.create({'name': 'Foo'})
+
+        with self.assertRaises(AttributeError):
+            instance.unknown = 42
